Add button above confidence meter to enable user to enter activity, organize related files
