+++ components/dashboard/Dashboard.js
@@ -15,7 +15,7 @@ import { isDone, mostRecent } from '../../src/app-helper';
 import { useAuth } from '../Auth';
 import { useSnackbar } from '../Snackbar';
 import ActivityCategoryTable from '../history/ActivityCategoryTable';
-import ActivityInputDialog from './ActivityInputDialog';
+import ActivityInputDialog from '../activityInput/ActivityInputDialog';
 import AirtablePropTypes from '../Airtable/PropTypes';
 import BackgroundHeader from '../BackgroundHeader';
 import EmploymentOutlookLauchpad from './EmploymentOutlookLauchpad';
